# frontend/openapi-v3.json
{"components":{"schemas":{"ErrorDetail":{"additionalProperties":false,"properties":{"location":{"description":"Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'","type":"string"},"message":{"description":"Error message text","type":"string"},"value":{"description":"The value at the given location"}},"type":"object"},"ErrorModel":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/ErrorModel.json"],"format":"uri","readOnly":true,"type":"string"},"detail":{"description":"A human-readable explanation specific to this occurrence of the problem.","examples":["Property foo is required but is missing."],"type":"string"},"errors":{"description":"Optional list of individual error details","items":{"$ref":"#/components/schemas/ErrorDetail"},"type":["array","null"]},"instance":{"description":"A URI reference that identifies the specific occurrence of the problem.","examples":["https://example.com/error-log/abc123"],"format":"uri","type":"string"},"status":{"description":"HTTP status code","examples":[400],"format":"int64","type":"integer"},"title":{"description":"A short, human-readable summary of the problem type. This value should not change between occurrences of the error.","examples":["Bad Request"],"type":"string"},"type":{"default":"about:blank","description":"A URI reference to human-readable documentation for the error.","examples":["https://example.com/errors/example"],"format":"uri","type":"string"}},"type":"object"},"Link":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/Link.json"],"format":"uri","readOnly":true,"type":"string"},"color":{"type":"string"},"icon":{"type":"string"},"id":{"type":"string"},"link":{"type":"string"},"sectionId":{"type":"string"},"title":{"type":"string"}},"required":["id","title","link","sectionId"],"type":"object"},"LinkBase":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/LinkBase.json"],"format":"uri","readOnly":true,"type":"string"},"color":{"type":"string"},"icon":{"type":"string"},"link":{"type":"string"},"sectionId":{"type":"string"},"title":{"type":"string"}},"required":["title","link","sectionId"],"type":"object"},"PublicShelf":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/PublicShelf.json"],"format":"uri","readOnly":true,"type":"string"},"description":{"type":"string"},"icon":{"type":"string"},"id":{"type":"string"},"path":{"type":"string"},"title":{"type":"string"}},"required":["id","title","description","icon","path"],"type":"object"},"Section":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/Section.json"],"format":"uri","readOnly":true,"type":"string"},"id":{"type":"string"},"shelfId":{"type":"string"},"title":{"type":"string"}},"required":["id","title","shelfId"],"type":"object"},"SectionBase":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/SectionBase.json"],"format":"uri","readOnly":true,"type":"string"},"shelfId":{"type":"string"},"title":{"type":"string"}},"required":["title","shelfId"],"type":"object"},"Setting":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/Setting.json"],"format":"uri","readOnly":true,"type":"string"},"key":{"type":"string"},"language_code":{"type":"string"},"value":{"type":"string"}},"required":["key","language_code","value"],"type":"object"},"SettingPageBody":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/SettingPageBody.json"],"format":"uri","readOnly":true,"type":"string"},"about":{"type":"string"},"about_show":{"type":"boolean"},"contact":{"type":"string"},"contact_show":{"type":"boolean"},"imprint":{"type":"string"},"imprint_show":{"type":"boolean"},"login_options":{"items":{"type":"string"},"type":["array","null"]},"privacy_policy":{"type":"string"},"privacy_policy_show":{"type":"boolean"},"redirect_to_dashboard":{"type":"boolean"},"terms_of_use":{"type":"string"},"terms_of_use_show":{"type":"boolean"}},"required":["about_show","about","contact_show","contact","imprint_show","imprint","terms_of_use_show","terms_of_use","privacy_policy_show","privacy_policy","redirect_to_dashboard","login_options"],"type":"object"},"Shelf":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/Shelf.json"],"format":"uri","readOnly":true,"type":"string"},"description":{"type":"string"},"domain":{"type":"string"},"icon":{"type":"string"},"id":{"type":"string"},"path":{"type":"string"},"theme":{"type":"string"},"title":{"type":"string"},"userId":{"type":"string"}},"required":["domain","theme","userId","id","title","description","icon","path"],"type":"object"},"ShelfBase":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/ShelfBase.json"],"format":"uri","readOnly":true,"type":"string"},"description":{"type":"string"},"domain":{"type":"string"},"icon":{"type":"string"},"path":{"pattern":"^[a-zA-Z0-9-]*$","patternDescription":"letters, numbers, and hyphens only","type":"string"},"theme":{"type":"string"},"title":{"minLength":1,"type":"string"},"userId":{"type":"string"}},"required":["title","path","userId"],"type":"object"},"Statistic":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/Statistic.json"],"format":"uri","readOnly":true,"type":"string"},"link_number":{"format":"int64","type":"integer"},"section_number":{"format":"int64","type":"integer"},"shelf_number":{"format":"int64","type":"integer"}},"required":["shelf_number","section_number","link_number"],"type":"object"},"User":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/User.json"],"format":"uri","readOnly":true,"type":"string"},"email":{"type":"string"},"first_name":{"type":"string"},"id":{"type":"string"},"last_name":{"type":"string"}},"required":["id","email","first_name","last_name"],"type":"object"},"UserBase":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/UserBase.json"],"format":"uri","readOnly":true,"type":"string"},"email":{"type":"string"},"first_name":{"type":"string"},"last_name":{"type":"string"}},"required":["email","first_name","last_name"],"type":"object"},"UserCreate":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/UserCreate.json"],"format":"uri","readOnly":true,"type":"string"},"email":{"type":"string"},"first_name":{"type":"string"},"last_name":{"type":"string"},"password":{"type":"string"}},"required":["password","email","first_name","last_name"],"type":"object"},"UserRequestBodyOnlyPassword":{"additionalProperties":false,"properties":{"$schema":{"description":"A URL to the JSON Schema for this object.","examples":["http://localhost:8085/schemas/UserRequestBodyOnlyPassword.json"],"format":"uri","readOnly":true,"type":"string"},"new_password":{"type":"string"},"old_password":{"type":"string"}},"required":["old_password","new_password"],"type":"object"}},"securitySchemes":{"LinkShelf":{"flows":{"authorizationCode":{"authorizationUrl":"/oauth/v2/authorize","refreshUrl":"/oauth/v2/token","scopes":{"email":"To return the email address.","offline_access":"?","openid":"To return the openid basic information.","profile":"To return the profile attributes like name."},"tokenUrl":"/oauth/v2/token"}},"type":"oauth2"}}},"info":{"description":"LinkShelf is an OpenSource alternative to linktree.","title":"LinkShelf","version":""},"openapi":"3.1.0","paths":{"/v1/links":{"get":{"description":"Get links by shelf ID.","operationId":"get-links","parameters":[{"explode":false,"in":"query","name":"shelfId","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Link"},"type":["array","null"]}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Get links by shelf ID","tags":["Link"]},"post":{"description":"Create a new link.","operationId":"post-create-link","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LinkBase"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Link"}}},"description":"Created"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Create link","tags":["Link"]}},"/v1/links/{linkId}":{"delete":{"description":"Delete a link by ID.","operationId":"delete-link","parameters":[{"explode":false,"in":"query","name":"shelfId","schema":{"type":"string"}},{"in":"path","name":"linkId","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Delete link","tags":["Link"]},"put":{"description":"Update an existing link.","operationId":"put-update-link","parameters":[{"explode":false,"in":"query","name":"shelfId","schema":{"type":"string"}},{"in":"path","name":"linkId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LinkBase"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Link"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Update link","tags":["Link"]}},"/v1/sections":{"get":{"description":"Get sections by shelf ID.","operationId":"get-sections","parameters":[{"explode":false,"in":"query","name":"shelfId","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Section"},"type":["array","null"]}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Get sections by shelf ID","tags":["Section"]},"post":{"description":"Create a new section.","operationId":"post-create-section","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionBase"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Section"}}},"description":"Created"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Create section","tags":["Section"]}},"/v1/sections/{sectionId}":{"delete":{"description":"Delete a section by ID.","operationId":"delete-section","parameters":[{"in":"path","name":"sectionId","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Delete section","tags":["Section"]},"put":{"description":"Update an existing section.","operationId":"put-update-section","parameters":[{"in":"path","name":"sectionId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionBase"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Section"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Update section","tags":["Section"]}},"/v1/settings":{"get":{"description":"Get page settings by language code.","operationId":"get-page-settings","parameters":[{"explode":false,"in":"query","name":"language_code","schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingPageBody"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Get page settings","tags":["Setting"]},"put":{"description":"Update page settings.","operationId":"put-update-setting","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Setting"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettingPageBody"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Update setting","tags":["Setting"]}},"/v1/shelves":{"get":{"description":"List all shelves.","operationId":"list-shelves","responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Shelf"},"type":["array","null"]}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"List shelves","tags":["Shelf"]},"post":{"description":"Create a new shelf.","operationId":"post-create-shelf","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShelfBase"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Shelf"}}},"description":"Created"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Create shelf","tags":["Shelf"]}},"/v1/shelves/by-path/{path}":{"get":{"description":"Get the public-safe view of a shelf by its path. Used to render a shelf's public link page and requires no authentication.","operationId":"get-public-shelf-by-path","parameters":[{"in":"path","name":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicShelf"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Get public shelf by path","tags":["Shelf"]}},"/v1/shelves/{shelfId}":{"delete":{"description":"Delete a shelf by ID.","operationId":"delete-shelf","parameters":[{"in":"path","name":"shelfId","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Delete shelf","tags":["Shelf"]},"get":{"description":"Get a shelf by ID.","operationId":"get-shelf-by-id","parameters":[{"in":"path","name":"shelfId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Shelf"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Get shelf by ID","tags":["Shelf"]},"put":{"description":"Update an existing shelf.","operationId":"put-update-shelf","parameters":[{"in":"path","name":"shelfId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShelfBase"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Shelf"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Update shelf","tags":["Shelf"]}},"/v1/statistics":{"get":{"description":"Get usage statistics for the current user.","operationId":"get-statistic","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Statistic"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Get statistic","tags":["Statistic"]}},"/v1/users":{"get":{"description":"List all users.","operationId":"list-users","responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/User"},"type":["array","null"]}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"List users","tags":["User"]},"post":{"description":"Create a new user.","operationId":"post-create-user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"description":"Created"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Create user","tags":["User"]}},"/v1/users/{userId}":{"delete":{"description":"Delete a user by ID.","operationId":"delete-user","parameters":[{"description":"The identifier of the chosen form you want.","in":"path","name":"userId","required":true,"schema":{"description":"The identifier of the chosen form you want.","type":"string"}}],"responses":{"204":{"description":"No Content"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Delete user","tags":["User"]},"get":{"description":"Get a user by ID.","operationId":"get-user-by-id","parameters":[{"description":"The identifier of the chosen form you want.","in":"path","name":"userId","required":true,"schema":{"description":"The identifier of the chosen form you want.","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Get user by ID","tags":["User"]},"put":{"description":"Update an existing user. Consider that password updates are not handled here.","operationId":"put-update-user","parameters":[{"description":"The identifier of the chosen form you want.","in":"path","name":"userId","required":true,"schema":{"description":"The identifier of the chosen form you want.","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}},"description":"OK"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Update user","tags":["User"]}},"/v1/users/{userId}/password":{"patch":{"description":"Patch a user's password.","operationId":"patch-user-password","parameters":[{"description":"The identifier of the chosen form you want.","in":"path","name":"userId","required":true,"schema":{"description":"The identifier of the chosen form you want.","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestBodyOnlyPassword"}}},"required":true},"responses":{"204":{"description":"No Content"},"default":{"content":{"application/problem+json":{"schema":{"$ref":"#/components/schemas/ErrorModel"}}},"description":"Error"}},"summary":"Patch user password","tags":["User"]}}},"servers":[{"url":"http://localhost:8085"},{"description":"This is the default server of LinkShelf","url":""}]}